+++ model/schema_delete_test.go
@@ -1,0 +1,10 @@
+package model
+
+import "testing"
+
+func TestDelete(t *testing.T) {
+	// TODO: Schema.Delete
+	// TODO: Schema.DeleteAt
+	// TODO: Schema.DeleteTx
+	// TODO: Schema.DeleteTxAt
+}